+++ src/com/actelion/research/gui/form/JSVGDetailView.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Actelion Pharmaceuticals Ltd., Gewerbestrasse 16, CH-4123 Allschwil, Switzerland
+ * Copyright 2017 Idorsia Pharmaceuticals Ltd., Hegenheimermattweg 91, CH-4123 Allschwil, Switzerland
  *
  * This file is part of DataWarrior.
  * 
@@ -18,24 +18,8 @@
 
 package com.actelion.research.gui.form;
 
-import java.awt.Component;
-import java.awt.Dimension;
-import java.awt.Frame;
-import java.awt.Graphics;
-import java.awt.Graphics2D;
-import java.awt.Image;
-import java.awt.Rectangle;
-import java.awt.Toolkit;
-import java.awt.datatransfer.StringSelection;
-import java.awt.event.ActionEvent;
-import java.awt.geom.Rectangle2D;
-import java.awt.print.PageFormat;
-import java.awt.print.Paper;
-import java.io.ByteArrayInputStream;
-import java.io.InputStream;
-
-import javax.swing.border.Border;
-
+import com.actelion.research.gui.JComponentCopyDialog;
+import com.actelion.research.util.Platform;
 import org.apache.batik.dom.svg.SAXSVGDocumentFactory;
 import org.apache.batik.swing.JSVGCanvas;
 import org.apache.batik.transcoder.TranscoderException;
@@ -44,8 +28,15 @@ import org.apache.batik.transcoder.print.PrintTranscoder;
 import org.apache.batik.util.XMLResourceDescriptor;
 import org.w3c.dom.svg.SVGDocument;
 
-import com.actelion.research.gui.JComponentCopyDialog;
-import com.actelion.research.util.Platform;
+import javax.swing.border.Border;
+import java.awt.*;
+import java.awt.datatransfer.StringSelection;
+import java.awt.event.ActionEvent;
+import java.awt.geom.Rectangle2D;
+import java.awt.print.PageFormat;
+import java.awt.print.Paper;
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
 
 public class JSVGDetailView extends JResultDetailView {
     private static final long serialVersionUID = 0x20070509;
@@ -193,7 +184,7 @@ public class JSVGDetailView extends JResultDetailView {
         }
 
     @Override
-	public void print(Graphics g, Rectangle2D.Float r, float scale, Object data) {
+	public void print(Graphics g, Rectangle2D.Double r, float scale, Object data) {
         try {
             String parser = XMLResourceDescriptor.getXMLParserClassName();
             SAXSVGDocumentFactory f = new SAXSVGDocumentFactory(parser);
